requirements.txt
django-extensions
django-debug-toolbar
django-braces

